chore: Consolidate test helper classes
