+++ src/Uno.UI.RuntimeTests/Helpers/FeatureConfigurationHelper.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Uno.Disposables;
 using Windows.UI.Xaml;
 
@@ -30,6 +31,20 @@ namespace Uno.UI.RuntimeTests.Helpers
 			return Disposable.Create(() => FeatureConfiguration.NativeListViewBase.RemoveItemAnimator = originalSetting);
 #else
 			return null;
+#endif
+		}
+
+		/// <summary>
+		/// On Android, ensure that managed popups are used for the duration of the test. On other platforms this is a no-op.
+		/// </summary>
+		public static IDisposable UseManagedPopups()
+		{
+#if !__ANDROID__
+			return null;
+#else
+			Assert.IsTrue(FeatureConfiguration.Popup.UseNativePopup); // If/when the default changes in SamplesApp, supplementary tests should be modified to test the new non-default
+			FeatureConfiguration.Popup.UseNativePopup = false;
+			return Disposable.Create(() => FeatureConfiguration.Popup.UseNativePopup = true);
 #endif
 		}
 	}
